src/sr_sac/requirements.txt
absl-py==1.4.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
brax==0.0.16
build==0.10.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.82
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
dataclasses==0.6
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distrax==0.1.4
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
entrypoints==0.4
etils==1.4.1
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.18
fastjsonschema==2.18.0
filelock==3.12.3
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
flax==0.7.2
fonttools==4.42.1
fqdn==1.5.1
future==0.18.3
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
glfw==2.6.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
gym==0.20.0
gym-notices==0.0.8
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
# jax==0.4.14
# jaxlib==0.4.14+cuda11.cudnn86
jaxopt==0.8
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
labmaze==1.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-collections==0.1.1
ml-dtypes==0.2.0
msgpack==1.0.5
mujoco==2.3.7
mujoco-py==2.1.2.14
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
numpy==1.25.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.3.5
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip-tools==7.3.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytinyrenderer==0.0.14
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-probability==0.20.1
tensorstore==0.1.41
terminado==0.17.1
tfp-nightly==0.20.0.dev20230524
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trimesh==3.9.35
# typing_extensions==4.5.0
uri-template==1.3.0
urllib3==1.26.16
wandb==0.15.9
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.7
zipp==3.16.2